+++ apollo-ios-pagination/Sources/ApolloPagination/GraphQLQueryPager.swift
@@ -2,15 +2,15 @@ import Apollo
 import ApolloAPI
 import Combine
 import Dispatch
+import Foundation
 
 /// Type-erases a query pager, transforming data from a generic type to a specific type, often a view model or array of view models.
-public class GraphQLQueryPager<Model> {
+public class GraphQLQueryPager<Model>: Publisher {
+  public typealias Failure = Never
   public typealias Output = Result<(Model, UpdateSource), Error>
-  private let _subject: CurrentValueSubject<Output?, Never> = .init(nil)
-
-  /// The `publisher` is the intended access point for using the pager as a `Combine` stream.
-  public var publisher: AnyPublisher<Output, Never> { _subject.compactMap { $0 }.eraseToAnyPublisher() }
-  public var cancellables = [AnyCancellable]()
+  let _subject: CurrentValueSubject<Output?, Never> = .init(nil)
+  var publisher: AnyPublisher<Output, Never> { _subject.compactMap { $0 }.eraseToAnyPublisher() }
+  public var cancellables: Set<AnyCancellable> = []
   public let pager: any PagerType
 
   public var canLoadNext: Bool { pager.canLoadNext }
@@ -137,6 +137,33 @@ public class GraphQLQueryPager<Model> {
   /// Resets pagination state and cancels further updates from the pager.
   public func cancel() {
     pager.cancel()
+  }
+
+  public func receive<S>(
+    subscriber: S
+  ) where S : Subscriber, Never == S.Failure, Result<(Model, UpdateSource), Error> == S.Input {
+    let subscription = PagerSubscription(pager: self, subscriber: subscriber)
+    subscriber.receive(subscription: subscription)
+  }
+}
+
+private class PagerSubscription<SubscriberType: Subscriber, Pager: GraphQLQueryPager<Model>, Model>: Subscription where SubscriberType.Input == Pager.Output {
+  private var subscriber: SubscriberType?
+  private var pager: Pager
+  private var cancellable: AnyCancellable?
+
+  init(pager: Pager, subscriber: SubscriberType) {
+    self.subscriber = subscriber
+    self.pager = pager
+    cancellable = pager.publisher.sink(receiveValue: {
+      _ = subscriber.receive($0)
+    })
+  }
+
+  func request(_ demand: Subscribers.Demand) { }
+
+  func cancel() {
+    subscriber = nil
   }
 }
